requirements.txt
langchain
requests
# ipykernel
# ollama
python-dotenv
langchain-community
langchain-core
langchain-groq
# pypdf
# bs4
# arxiv
# pymupdf
langchain-text-splitters
# langchain_chroma
sentence-transformers
cohere
langchain-huggingface
faiss-cpu
python-telegram-bot==13.15
# torch==2.0.1 # Optimization